created main menu button
if they win, then they will have buttons to let them go to the main menu, the game mode screen, or exit the game.
if they lose, they have buttons to go to the main menu, restart the level, or exit. Waiting for the main menu and game mode screens to be completed to add functionality to my buttons.